updated design a little
